Limit Xray validation to manual saves
